Use qint64 for ScopedTimer/SuspendableTimer/Timer to match the resolution of PerformanceTimer.

+++ src/util/timer.h
@@ -20,11 +20,11 @@ class Timer {
     // Restart the timer returning the nanoseconds since it was last
     // started/restarted. If report is true, reports the elapsed time to the
     // associated Stat key.
-    int restart(bool report);
+    qint64 restart(bool report);
 
     // Returns nanoseconds since start/restart was called. If report is true,
     // reports the elapsed time to the associated Stat key.
-    int elapsed(bool report);
+    qint64 elapsed(bool report);
 
   protected:
     QString m_key;
@@ -38,12 +38,12 @@ class SuspendableTimer : public Timer {
     SuspendableTimer(const QString& key,
             Stat::ComputeFlags compute = kDefaultComputeFlags);
     void start();
-    int suspend();
+    qint64 suspend();
     void go();
-    int elapsed(bool report);
+    qint64 elapsed(bool report);
 
   private:
-    int m_leapTime;
+    qint64 m_leapTime;
 };
 
 class ScopedTimer {
